% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FL_all}
\alias{FL_all}
\title{Threat and demography scores for 54 ESA-listed species in FL.}
\format{A data frame with 50 observations and 4 variables
\describe{
\item{\code{Species}}{The species that was scored}
\item{\code{Demography}}{The score for the demographic change status}
\item{\code{Threat}}{The score for the threat change status}
\item{\code{Recommendation}}{FWS status change recommendation from biennial report to Congress, 2012.}
}}
\source{
Defenders of Wildlife
}
\usage{
FL_all
}
\description{
We scored the threat and demographic changes for 54 ESA-listed, non-plant
species in Florida over which FWS has primary jurisdiction.
}
\keyword{datasets}

