% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_lda_comps}
\alias{get_lda_comps}
\title{Calculate consistency (and deviations thereof) for an LDA model.}
\usage{
get_lda_comps(dat, mod)
}
\arguments{
\item{dat}{The data on which the LDA was run}

\item{mod}{The results of an LDA model}
}
\value{
A list with:
\itemize{
 \item{marg_tab}{The margin table}
 \item{diag_marg_tab}{The diagonal of the margin table}
 \item{consistency}{The sum of the diagonal, for overall consistency}}
}
\description{
Tabulates the observed response variable and the classifications from the
LDA model to calculate proportions assigned correctly between classes and
overall consistency of assignments.
}
\seealso{
MASS::lda
}

