% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_multinom_p}
\alias{get_multinom_p}
\title{Calculate the p-values for an nnet::multinomial model.}
\usage{
get_multinom_p(mod)
}
\arguments{
\item{mod}{The results for a multinom model}
}
\value{
A matrix of p-values for predictors and response categories.
}
\description{
Uses pnorm of the zscore to return approximate p-values for model mod.
}
\seealso{
nnet::multinomial
}

