% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lda_class}
\alias{lda_class}
\title{Data from LDA classification of FWS status changes and threat & demography scores.}
\format{A data frame with 36 observations and 6 variables
\describe{
\item{\code{Model}}{Which of four models gave these results}
\item{\code{classify}}{The LDA-based classification}
\item{\code{FWS_rec}}{FWS's status change recommendation}
\item{\code{Pct_cross}}{The percentage of species in this combo}
\item{\code{N_cross}}{The number of species in this combo}
\item{\code{col}}{The color that should be used in the plot}
}}
\source{
data source, if any
}
\usage{
lda_class
}
\description{
The discriminant function analyses that gave rise to this data are in the
Analysis 1 vignette.
}
\keyword{datasets}

