% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi}
\alias{multi}
\title{Threat and demography scores from multiple scorers.}
\format{A data frame with 54 observations and 4 variables
\describe{
\item{\code{Person}}{Individual who generated the scores}
\item{\code{Species}}{The species that was scored}
\item{\code{Threat}}{The score for the threat change status}
\item{\code{Demography}}{The score for the demographic change status}
}}
\source{
Defenders of Wildlife
}
\usage{
multi
}
\description{
To evaluate the consistency of scoring threat and demographic status from
multiple people, five individuals used 5-yr reviews and Fed. Reg. documents
to derive scores for ten randomly selected FL species using a scoring key.
}
\keyword{datasets}

