% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{multiplot}
\alias{multiplot}
\title{Create multi-panel plot from individual plots.}
\source{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{One or more plot objects to be plotted}

\item{plotlist}{A list of plots, if not passed as ...}

\item{cols}{Number of columns the multi-panel plot will have.}

\item{layout}{An option matrix describing the layout of the multi-panel plot.}
}
\value{
A multi-panel plot composed of the input plots.
}
\description{
A simplistic way to create multi-panel ggplot figures when facet won't work.
}
\examples{
# plot_a <- qplot(data = dat, geom = "hist")
# plot_b <- qplot(data = dat2, x = V1, y = V2, geom = "jitter")
# multiplot(plot_a, plot_b, cols = 2)
}
\seealso{
The gtable package \url{https://github.com/hadley/gtable} has much
more advanced multiplot layout capabilities, but is more complex.
}

